+++ python/dalex/dalex/fairness/_basics/_base_objects.py
@@ -1,6 +1,5 @@
 import abc
-from .exceptions import ParameterCheckError
-from .checks import check_parameters
+from . import checks
 
 
 class _AbsObject(metaclass=abc.ABCMeta):
@@ -15,7 +14,7 @@ class _AbsObject(metaclass=abc.ABCMeta):
 class _FairnessObject(_AbsObject):
 
     def __init__(self, y, y_hat, protected, privileged, verbose):
-        y, y_hat, protected, privileged = check_parameters(y, y_hat, protected, privileged, verbose)
+        y, y_hat, protected, privileged = checks.check_parameters(y, y_hat, protected, privileged, verbose)
         self.privileged = privileged
         self.protected = protected
         self.y_hat = y_hat