+++ python/dalex/dalex/predict_explanations/_ceteris_paribus/object.py
@@ -9,30 +9,30 @@ from ... import _theme, _global_checks, _global_utils
 
 
 class CeterisParibus:
-    """Calculate instance level variable profiles as Ceteris Paribus
+    """Calculate predict-level variable profiles as Ceteris Paribus
 
     Parameters
     -----------
     variables : array_like of str, optional
         Variables for which the profiles will be calculated
-        (default is None, which means all of the variables).
+        (default is `None`, which means all of the variables).
     grid_points : int, optional
-        Maximum number of points for profile calculations (default is 101).
+        Maximum number of points for profile calculations (default is `101`).
         NOTE: The final number of points may be lower than `grid_points`,
         eg. if there is not enough unique values for a given variable.
     variable_splits : dict of lists, optional
-        Split points for variables e.g. {'x': [0, 0.2, 0.5, 0.8, 1], 'y': ['a', 'b']}
-        (default is None, which means that they will be calculated using one of
+        Split points for variables e.g. `{'x': [0, 0.2, 0.5, 0.8, 1], 'y': ['a', 'b']}`
+        (default is `None`, which means that they will be calculated using one of
         `variable_splits_type` and the `data` attribute).
     variable_splits_type : {'uniform', 'quantiles'}, optional
-        Way of calculating `variable_splits`. Set 'quantiles' for percentiles.
-        (default is 'uniform', which means uniform grid of points).
+        Way of calculating `variable_splits`. Set `'quantiles'` for percentiles.
+        (default is `'uniform'`, which means uniform grid of points).
     variable_splits_with_obs: bool, optional
         Add variable values of `new_observation` data to the `variable_splits`
-        (default is True).
+        (default is `True`).
     processes : int, optional
         Number of parallel processes to use in calculations. Iterated over `variables`
-        (default is 1, which means no parallel computation).
+        (default is `1`, which means no parallel computation).
 
     Attributes
     -----------
@@ -99,7 +99,7 @@ class CeterisParibus:
         y : pd.Series or np.ndarray (1d), optional
             Target variable with the same length as `new_observation`.
         verbose : bool, optional
-            Print tqdm progress bar (default is True).
+            Print tqdm progress bar (default is `True`).
 
         Returns
         -----------
@@ -150,33 +150,36 @@ class CeterisParibus:
         Parameters
         -----------
         objects : CeterisParibus object or array_like of CeterisParibus objects
-            Additional objects to plot in subplots (default is None).
+            Additional objects to plot in subplots (default is `None`).
         variable_type : {'numerical', 'categorical'}
-            Plot the profiles for numerical or categorical variables (default is 'numerical').
+            Plot the profiles for numerical or categorical variables 
+            (default is `'numerical'`).
         variables : str or array_like of str, optional
             Variables for which the profiles will be calculated
-            (default is None, which means all of the variables).
+            (default is `None`, which means all of the variables).
         size : float, optional
-            Width of lines in px (default is 2).
+            Width of lines in px (default is `2`).
         alpha : float <0, 1>, optional
-            Opacity of lines (default is 1).
+            Opacity of lines (default is `1`).
         color : str, optional
-            Variable name used for grouping (default is '_label_', which groups by models).
+            Variable name used for grouping
+            (default is `'_label_'`, which groups by models).
         facet_ncol : int, optional
-            Number of columns on the plot grid (default is 2).
+            Number of columns on the plot grid (default is `2`).
         show_observations : bool, optional
-            Show observation points (default is True).
+            Show observation points (default is `True`).
         title : str, optional
-            Title of the plot (default is "Ceteris Paribus Profiles").
+            Title of the plot (default is `"Ceteris Paribus Profiles"`).
         y_title : str, optional
-            Title of the y/x axis (default is "prediction").
+            Title of the y/x axis (default is `"prediction"`).
         horizontal_spacing : float <0, 1>, optional
-            Ratio of horizontal space between the plots (default depends on `variable_type`).
+            Ratio of horizontal space between the plots
+            (default depends on `variable_type`).
         vertical_spacing : float <0, 1>, optional
-            Ratio of vertical space between the plots (default is 0.3/number of rows).
+            Ratio of vertical space between the plots (default is `0.3/number of rows`).
         show : bool, optional
-            True shows the plot; False returns the plotly Figure object that can be
-            edited or saved using the `write_image()` method (default is True).
+            `True` shows the plot; `False` returns the plotly Figure object that can 
+            be edited or saved using the `write_image()` method (default is `True`).
 
         Returns
         -----------